server/requirements.txt
django==5.0.3
django-appconf
django-cors-headers 
djangorestframework
PyJWT